Margins, new buttons, new page
Made a new page to talk about opening the Terminal first, might be useful.
Edited the text within the intro to be more fluid and accurate, added sections ready for image of Linus and RMS.
Added new buttons on Lessons page with buttons that lead to coming soon, will make adding pages to those buttons quicker later on as it'll be one line to edit.